api/requirement.txt
absl-py==0.8.1
albumentations==0.4.3
alembic==1.4.2
aniso8601==8.0.0
astor==0.8.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bcrypt==3.2.0
bleach==3.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
conda==4.3.16
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.31.0
h5py==2.10.0
idna==2.10
imageio==2.6.1
imgaug==0.2.6
importlib-metadata==1.1.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kaggle==1.5.6
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
# mkl-fft==1.0.15
# mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
mtcnn==0.1.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
opencv-contrib-python==4.1.2.30
opencv-python==4.2.0.32
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
pkgconfig==1.5.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.23.1
scipy
Send2Trash==1.5.0
six==1.13.0
SQLAlchemy==1.3.19
tensorboard
tensorboard-plugin-wit
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tornado==6.0.3
tqdm==4.46.0
traitlets==4.3.3
urllib3
virtualenv==16.7.9
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
